crop image, ahora las imagenes quedan siempre bien, correccion quitar canciones de una playlist, y mejora del diseño de BibliotecaFragment

+++ app/src/main/java/com/example/reproductormultimedia_vj/Adapter/AddCancionPlaylistAdapter.java
@@ -88,7 +88,7 @@ public class AddCancionPlaylistAdapter extends RecyclerView.Adapter<AddCancionPl
             if(check.isChecked()){
                 AddPlaylistActivity.playlist.getListaCanciones().add(cancion.getIdCancion());
             }else{
-                AddPlaylistActivity.playlist.getListaCanciones().remove(cancion.getIdCancion());
+                AddPlaylistActivity.playlist.getListaCanciones().remove((Integer)cancion.getIdCancion());
             }
         }
     }
